Add validation for NewExpense pages, add Atom to split details
