+++ src/common/_buttons.scss
@@ -9,7 +9,7 @@
   font-family: 'Inter', sans-serif;
   font-weight: 600;
   font-size: 1rem;
-  border: 1px solid white;
+  border: none;
   cursor: pointer;
   transition: width 0.5s ease-in-out;
   &:active {
@@ -83,10 +83,6 @@
     color: v.$dark-gray;
     border: 1px solid v.$dark-gray;
     cursor: not-allowed;
-  }
-
-  &.button--border-none {
-    border: none;
   }
 }
