+++ src/common/_form.scss
@@ -13,11 +13,9 @@
 }
 
 .input-label {
-  font-family: 'Rubik', sans-serif;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: v.$dark-gray;
-  text-transform: uppercase;
+  font-size: 1rem;
+  font-weight: 600;
+  color: v.$blue-black;
 }
 
 .text-input {
