+++ src/pages/NewExpense/NewExpense.scss
@@ -23,6 +23,7 @@
   }
 
   .expense-input-container {
+    position: relative;
     z-index: 3;
     display: grid;
     gap: 1.5rem;
@@ -66,5 +67,16 @@
     select {
       color: v.$washed-gray;
     }
+  }
+}
+
+.split-details-page {
+  .nucleus__amount {
+    font-family: 'Rubik', sans-serif;
+    font-size: 1.5rem;
+    font-weight: 500;
+    color: #ffffff;
+    letter-spacing: 0.02em;
+    text-shadow: 0px 2px 0px rgba(0, 0, 0, 0.15);
   }
 }
